update deps, hotfix for supply (#15)

+++ sources/libs/CoinHelper.move
@@ -51,7 +51,7 @@ module MultiSwap::CoinHelper {
     /// Get supply for `CoinType`.
     /// Would throw error if supply for `CoinType` doesn't exist.
     public fun supply<CoinType>(): u64 {
-        Option::extract(&mut Coin::supply<CoinType>())
+        (Option::extract(&mut Coin::supply<CoinType>()) as u64)
     }
 
     /// Generate LP coin name for pair `X`/`Y`.
